requirements.txt
Flask==2.1.2
flask-login==0.5.0
werkzeug==2.0.3
psycopg2
twilio==9.4.1
python-dotenv==1.0.0
gunicorn==20.1.0
APScheduler==3.9.1.post1
